% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seu.R
\name{check_seu}
\alias{check_seu}
\title{check_seu: checking if mitoRatio (or any colnumn) is present in your Seurat object}
\usage{
check_seu(seurat_obj, check = "mitoRatio")
}
\arguments{
\item{seurat_obj}{Provide your Seurat object}

\item{check}{Provide a column name \link{default: 'mitoRatio'}}
}
\value{
Checked Seurat object
}
\description{
check_seu: checking if mitoRatio (or any colnumn) is present in your Seurat object
}
\examples{
# DO NOT RUN 
# devtools::install_github('satijalab/seurat-data')
# library(SeuratData)
# AvailableData()
# InstallData("pbmc3k")
# data("pbmc3k")
# pbmc3k$mitoRatio = PercentageFeatureSet(object = pbmc3k,pattern = "^MT-")/100
# check_seu(pbmc3k,"mitoRatio")
# DO NOT RUN


}
\seealso{
Use this function before performing mito-cutoff change point detection with SCdetMito.
}
\keyword{internal}
