% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCdetMito.R
\name{SCdetMito}
\alias{SCdetMito}
\title{SCdetMito: Do mitoRatio change point detection}
\usage{
SCdetMito(
  seurat_obj,
  mitoRatio = "mitoRatio",
  by = "sample",
  bin_width = 0.01,
  min_cut = 0.01,
  max_cut = 1,
  table_out = FALSE,
  plot = TRUE,
  diff_num = 5,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data}

\item{mitoRatio}{The column name for mitoRatio in the Seurat object}

\item{by}{The target variable for neighbor comparison would be the 'samples' column, which would be used to conduct counts distribution and tests for each sample. The result would be a mitoRation cutoff for those sample. The target variable colnumn could be 'group' or 'treatment' either.}

\item{bin_width}{0.01(default), slide window width between every two cutoff, any value less than 0.1 could be used}

\item{min_cut}{0.01(default), minimum cutoff during the detection process, any value less than 0.1 could be used}

\item{max_cut}{1(default), maximum cutoff during the detection process, any value less than 1 could be used}

\item{table_out}{Set to TRUE or FALSE(default) for table generation}

\item{plot}{Set to TRUE or FALSE(default) for plots generation}

\item{diff_num}{Cell changed-counts, which at least filtered by a slide cut off}

\item{...}{Additional parameters}
}
\value{
mitoRatio cutoff value and (optional) some processed tables and plots
}
\description{
SCdetMito: Do mitoRatio change point detection
}
\examples{
# DO NOT RUN
# devtools::install_github('satijalab/seurat-data')
# library(SeuratData)
# AvailableData()
# InstallData("pbmc3k")
# data("pbmc3k")
# pbmc3k
# library(Seurat)
# pbmc3k$mitoRatio = PercentageFeatureSet(object = pbmc3k,
#                                         pattern = "^MT-")/100
# head(pbmc3k)
#
# library(SCdetMito)
#
# take "orig.ident" as one sample
# mitoRatio_cut_off = SCdetMito(pbmc3k,
#                                by="orig.ident",
#                                table_out = T)
#
#
# generate 3 samples for test
# dim(pbmc3k)
# pbmc3k$samples = rep(c("A","B","C"),
#                      c(900,900,900))
# mitoRatio_cut_off = SCdetMito(pbmc3k,
#                               by="samples",
#                               table_out = T)
# DO NOT RUN


}
\seealso{
Use this function to identify mitoRatio change points in your single-cell RNA-seq data.
}
\keyword{internal}
