% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCQC_processedPlots.R
\name{SCQC_processedPlots}
\alias{SCQC_processedPlots}
\title{SCQC_processedPlots: Draw processed plots}
\usage{
SCQC_processedPlots(seurat_obj, by, flag)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data}

\item{by}{Sample/group columns}

\item{flag}{Flag for out put files}
}
\value{
processed plots
}
\description{
SCQC_processedPlots: Draw processed plots
}
\examples{
# DO NOT RUN
# devtools::install_github('satijalab/seurat-data')
# library(SeuratData)
# AvailableData()
# InstallData("pbmc3k")
# data("pbmc3k")
# head(pbmc3k)
# library(Seurat)
# pbmc3k$mitoRatio <- PercentageFeatureSet(object = pbmc3k, pattern = "^MT-") / 100
# max(pbmc3k$mitoRatio)
# min(pbmc3k$mitoRatio)
# library(SCdetMito)
# dim(pbmc3k)
# pbmc3k$samples <- rep(
#    c("A", "B", "C"),
#    c(900, 900, 900)
#  )
# Idents(pbmc3k) <- pbmc3k$samples
# pbmc3k_A <- subset(pbmc3k,
#     idents = "A"
#  )
# qc： optional-1
#  qcpassed_pbmc3k_A <- SCQCone(pbmc3k_A,
#      max_mito = 0.06,
#      removeDouble = FALSE
#  )
#  dim(qcpassed_pbmc3k_A)
# qc： optional-2
#  qcpassed_pbmc3k_A <- SCQCone(pbmc3k_A,
#      max_mito = 0.06,
#      removeDouble = FALSE
#  )
#  dim(qcpassed_pbmc3k_A)
# qc： optional-3
#  qcpassed_pbmc3k_A <- SCQCone(pbmc3k_A,
#      max_mito = "SCdetMito",
#      removeDouble = FALSE
#  )
#  dim(qcpassed_pbmc3k_A)

}
\seealso{
Use this function to generate processed plots
}
\keyword{internal}
