% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCQCone.R
\name{SCQCone}
\alias{SCQCone}
\title{SCQCone: Perform quality control (QC) for one sample/group single-cell RNA-seq data}
\usage{
SCQCone(
  seurat_obj,
  nFeature_RNA = "nFeature_RNA",
  nCount_RNA = "nCount_RNA",
  mitoRatio = "mitoRatio",
  min_genes = 200,
  max_genes = Inf,
  min_counts = 500,
  max_counts = Inf,
  min_mito = 0,
  max_mito = "SCdetMito",
  removeDouble = TRUE,
  plot = TRUE,
  table_out = TRUE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data for a single sample}

\item{min_genes}{Minimum number of genes expressed to consider a cell \link{default: 200}}

\item{max_genes}{Maximum number of genes expressed to consider a cell \link{default: Inf}}

\item{min_counts}{Minimum total counts per cell to consider a cell \link{default: 500}}

\item{max_counts}{Maximum total counts per cell to consider a cell \link{default: Inf}}

\item{min_mito}{Minimum percentage of mitochondrial genes per cell to consider a cell \link{default: 0}}

\item{max_mito}{Maximum percentage of mitochondrial genes per cell to consider a cell \link{default: 20}}

\item{removeDouble}{Do doublets cell or not \link{default: TRUE}}

\item{plot}{Set to TRUE to generate QC plots \link{default: TRUE}}

\item{table_out}{Set to TRUE to generate QC summary table \link{default: TRUE}}

\item{...}{Additional parameters (not used)}
}
\value{
QC summarys, QC-passed seurat object, and, if specified, QC plots
}
\description{
This function performs QC for a single-cell RNA-seq sample/group. It includes steps such as filtering cells based on library size,
mitochondrial content, and other relevant metrics. It generates various QC plots and tables to assess the quality of the data.
}
\examples{
# DO NOT RUN
# devtools::install_github('satijalab/seurat-data')
# library(SeuratData)
# AvailableData()
# InstallData("pbmc3k")
# data("pbmc3k")
# pbmc3k$mitoRatio = PercentageFeatureSet(object = pbmc3k, pattern = "^MT-")/100
# seurat_obj <- SCQCone(pbmc3k)
# DO NOT RUN

}
\seealso{
Use this function to perform QC for a single-cell RNA-seq sample/group.
}
\keyword{internal}
